refactor: break libs plugin into additional files, and markdown urls plugins
