+++ docs/plugins/src/mod.ts
@@ -1,0 +1,3 @@
+throw new Error(
+  "There are no exports in the root module of this package. Import from submodules instead. Use intellisense or the docs to find the available submodules.",
+);
