+++ docs/plugins/src/utils/fs.ts
@@ -1,0 +1,10 @@
+import fs from "node:fs/promises";
+import path from "node:path";
+
+export async function writeFile(
+  filePath: string,
+  content: string,
+): Promise<void> {
+  await fs.mkdir(path.dirname(filePath), { recursive: true });
+  await fs.writeFile(filePath, content, "utf-8");
+}
