+++ docs/plugins/src/utils/types.ts
@@ -1,0 +1,6 @@
+import { docsSchema } from "@astrojs/starlight/schema";
+import type { z } from "astro/zod";
+
+export type Frontmatter = Partial<
+  z.infer<ReturnType<ReturnType<typeof docsSchema>>>
+>;
