--- docs/src/plugins/libs/index.ts
@@ -1,1 +1,0 @@
-export * from './libs-plugin';
