--- docs/src/plugins/libs/libs-plugin.ts
@@ -1,316 +1,0 @@
-import fs from 'node:fs/promises';
-import path from 'node:path';
-import { Application, ProjectReflection, ReflectionKind, type TypeDocOptions } from 'typedoc';
-
-import { type AstroIntegrationLogger, type RemarkPlugins } from 'astro';
-import type { StarlightPlugin } from '@astrojs/starlight/types';
-import { docsSchema } from '@astrojs/starlight/schema';
-import type { z } from 'astro/zod';
-import { type PluginOptions as TypeDocMarkdownOptions } from 'typedoc-plugin-markdown';
-import { visit } from 'unist-util-visit';
-import yaml from 'yaml';
-
-const CONTENT_DIR = path.resolve('src', 'content');
-const TMP_DIR = path.resolve(CONTENT_DIR, 'tmp');
-const DOCS_DIR = path.resolve(CONTENT_DIR, 'docs');
-
-export type Frontmatter = Partial<z.infer<ReturnType<ReturnType<typeof docsSchema>>>>;
-
-async function generateApiDocs({
-  baseDir,
-  typeDoc,
-}: LibsLoaderOptions): Promise<ProjectReflection> {
-  const defaultTypeDocOptions: LibsLoaderTypeDocOptions = {
-    entryPoints: [`${baseDir}/*`],
-    entryPointStrategy: 'packages',
-    packageOptions: {
-      entryPoints: ['src/index.ts'],
-      tsconfig: './tsconfig.json',
-      readme: 'none',
-    },
-    plugin: ['typedoc-plugin-markdown', 'typedoc-plugin-frontmatter'],
-    tsconfig: './tsconfig.typedoc.json',
-    outputs: [{ name: 'markdown', path: TMP_DIR }],
-    readme: 'none',
-    hidePageTitle: true,
-    hideBreadcrumbs: true,
-    hidePageHeader: true,
-  };
-
-  const app = await Application.bootstrapWithPlugins({
-    ...defaultTypeDocOptions,
-    ...typeDoc,
-  });
-
-  const project = await app.convert();
-  if (!project) {
-    throw new Error('Failed to convert project with TypeDoc');
-  }
-  await app.generateOutputs(project);
-
-  return project;
-}
-
-/**
- * Options for the LibsLoader plugin.
- */
-export interface LibsLoaderOptions {
-  /**
-   * Base directory where the libraries are located.
-   *
-   * This should be a path relative to your `astro.config.mjs` file.
-   */
-  baseDir: string | URL;
-
-  /**
-   * Output directory for the generated documentation.
-   *
-   * This should be a path relative to the `src/content/docs` directory.
-   * Defaults to `libs`.
-   */
-  outDir?: string | URL;
-
-  /**
-   * Whether to clean the output directory before generating documentation.
-   *
-   * Defaults to `true`.
-   */
-  clean?: boolean;
-
-  /**
-   * Options for TypeDoc.
-   */
-  typeDoc?: LibsLoaderTypeDocOptions;
-
-  /**
-   * Additional files to include in the documentation.
-   */
-  additionalFiles?: { path: string; frontmatter?: Frontmatter }[];
-
-  /**
-   * Frontmatter applied to every markdown file generated by TypeDoc.
-   */
-  frontmatter?: Frontmatter;
-}
-
-export type LibsLoaderTypeDocOptions = TypeDocMarkdownOptions & TypeDocOptions;
-
-type RemarkPlugin = RemarkPlugins[number];
-
-const prettyUrlsPlugin: RemarkPlugin =
-  ([logger, docsDir, site, baseUrl]: [AstroIntegrationLogger, string, string, string]) =>
-  async (tree, file) => {
-    const currentFileDir = path.dirname(file.path);
-
-    visit(tree, 'link', node => {
-      const url = node.url;
-
-      // take full URLs to the current site and make them relative
-      if (url.startsWith(site)) {
-        node.url = new URL(url).pathname;
-        return;
-      }
-
-      // skip any other full URLs
-      if (
-        url.startsWith('https://') ||
-        url.startsWith('/') ||
-        url.startsWith('http://') ||
-        url.startsWith('mailto:') ||
-        url.startsWith('#')
-      ) {
-        logger.debug(`Skipping URL: ${url}`);
-        return;
-      }
-
-      // normalize all other relative URLs to the docs directory
-      const absoluteLinkedFilePath = path.resolve(currentFileDir, url);
-      const relativeToDocs = path.relative(docsDir, absoluteLinkedFilePath);
-      const normalizedUrl = `${baseUrl}${relativeToDocs.replace(/(index)?\.mdx?(#.*)?$/, '$2').toLowerCase()}`;
-      logger.debug(`Normalizing URL: ${url} -> ${normalizedUrl}`);
-
-      node.url = normalizedUrl;
-    });
-  };
-
-export function libsPlugin(opts: LibsLoaderOptions): StarlightPlugin {
-  return {
-    name: 'libs-starlight-plugin',
-    hooks: {
-      async 'config:setup'(ctx) {
-        const site = ctx.astroConfig.site;
-        const baseDir = path.resolve(opts.baseDir.toString());
-        const outDir = path.resolve(DOCS_DIR, opts.outDir?.toString() ?? 'libs');
-        const clean = opts.clean ?? true;
-
-        if (!site) {
-          throw new Error('Site URL is not defined in Astro config');
-        }
-
-        if (clean) {
-          await fs.rm(outDir, { recursive: true, maxRetries: 3, force: true });
-          await fs.rm(TMP_DIR, { recursive: true, maxRetries: 3, force: true });
-        }
-
-        ctx.addIntegration({
-          name: 'libs-astro-plugin',
-          hooks: {
-            'astro:config:setup': ({ updateConfig, config, logger }) => {
-              updateConfig({
-                markdown: {
-                  remarkPlugins: [
-                    ...config.markdown.remarkPlugins,
-                    [prettyUrlsPlugin, [logger, DOCS_DIR, site, ctx.astroConfig.base]],
-                  ],
-                },
-              });
-            },
-          },
-        });
-
-        const project = await generateApiDocs(opts);
-
-        const librarySidebarItems = [];
-        const modules = project.getChildrenByKind(ReflectionKind.Module);
-        for (const { name } of modules) {
-          const id = name.startsWith('@') ? name.split('/')[1] : name;
-          const outputRootDir = path.resolve(outDir, id);
-          const outputApiDir = path.resolve(outputRootDir, 'api');
-          const title = titleFromIdCapitalized(id);
-
-          await processMarkdown({
-            inputPath: path.resolve(baseDir, id, 'README.md'),
-            outputPath: path.resolve(outputRootDir, `index.md`),
-            frontmatter: {
-              title,
-              ...(opts?.frontmatter || {}),
-            },
-          });
-
-          const apiSrcDir = path.resolve(TMP_DIR, name);
-          const files = await fs.readdir(apiSrcDir, {
-            withFileTypes: true,
-            recursive: true,
-          });
-          for (const file of files) {
-            if (file.isFile() && file.name.endsWith('.md')) {
-              const prefix = path.relative(apiSrcDir, file.parentPath);
-              const inputFileName = file.name;
-              const isReadme = inputFileName.endsWith('README.md');
-              const outputFileName = isReadme ? 'index.md' : inputFileName;
-              const title = isReadme ? 'Overview' : titleFromFilename(file.name);
-
-              await processMarkdown({
-                inputPath: path.resolve(apiSrcDir, prefix, inputFileName),
-                outputPath: path.resolve(outputApiDir, prefix, outputFileName),
-                frontmatter: {
-                  title,
-                  ...(opts?.frontmatter || {}),
-                },
-              });
-            }
-          }
-
-          librarySidebarItems.push({
-            label: title,
-            collapsed: true,
-            items: [
-              {
-                label: 'Overview',
-                link: `/libs/${id}`,
-              },
-              {
-                label: 'API Reference',
-                collapsed: true,
-                autogenerate: {
-                  collapsed: true,
-                  directory: `libs/${id}/api`,
-                },
-              },
-            ],
-          });
-        }
-
-        const sidebarItems = [];
-        for (const file of opts.additionalFiles || []) {
-          const fileName = path.basename(file.path).toLowerCase();
-          const id = idFromFilename(fileName);
-          const title = titleFromIdCapitalized(id);
-
-          await processMarkdown({
-            inputPath: path.resolve(file.path),
-            outputPath: path.resolve(DOCS_DIR, fileName),
-            frontmatter: {
-              title,
-              ...(file.frontmatter || {}),
-            },
-          });
-
-          sidebarItems.push({
-            label: title,
-            link: `/${id}`,
-          });
-        }
-
-        ctx.updateConfig({
-          sidebar: [
-            { label: 'Libraries', items: librarySidebarItems },
-            ...(ctx.config.sidebar || []),
-            ...sidebarItems,
-          ],
-        });
-      },
-    },
-  };
-}
-
-async function writeFile(filePath: string, content: string): Promise<void> {
-  await fs.mkdir(path.dirname(filePath), { recursive: true });
-  await fs.writeFile(filePath, content, 'utf-8');
-}
-
-function idFromFilename(fileName: string): string {
-  return fileName.replace(/\.mdx?$/, '');
-}
-
-function titleFromFilename(fileName: string): string {
-  return titleFromId(idFromFilename(fileName));
-}
-
-function titleFromId(id: string): string {
-  return id.replace(/-/g, ' ');
-}
-
-function titleFromIdCapitalized(id: string): string {
-  return titleFromId(id).replace(/\b\w/g, char => char.toUpperCase());
-}
-
-interface ProcessMarkdownOpts {
-  inputPath: string;
-  outputPath: string;
-  frontmatter: Frontmatter;
-}
-
-async function processMarkdown({
-  inputPath,
-  outputPath,
-  frontmatter,
-}: ProcessMarkdownOpts): Promise<void> {
-  const input = await fs.readFile(inputPath, 'utf-8');
-
-  const output = addFrontmatter(input, frontmatter)
-    .replace(/^\s*#\s*.*$/m, '')
-    .replaceAll('README.md', 'index.md');
-
-  await writeFile(outputPath, output);
-}
-
-function addFrontmatter(content: string, frontmatter: Frontmatter): string {
-  const frontmatterStr = yaml.stringify(frontmatter);
-
-  if (frontmatterStr.length === 0) {
-    return content;
-  }
-
-  return `---\n${frontmatterStr}---\n\n${content}`;
-}
