--- docs/src/plugins/theme/index.ts
@@ -1,1 +1,0 @@
-export * from './dfinity-starlight-theme';
