+++ pnpm-workspace.yaml
@@ -1,5 +1,6 @@
 packages:
   - docs
+  - docs/plugins
   - e2e/node
   - packages/agent
   - packages/assets
